LPS-98680 Only try to remove when the persistence test file exists (for other persistence* files the issue was fixed by 84cc8ab349afd7983436db27075bd6a78eb73ce7)

+++ modules/util/portal-tools-service-builder/src/main/java/com/liferay/portal/tools/service/builder/ServiceBuilder.java
@@ -3293,7 +3293,9 @@ public class ServiceBuilder {
 				_testOutputPath, "/service/persistence/test/", entity.getName(),
 				"PersistenceTest.java"));
 
-		if (entity.isDeprecated() || !entity.hasPersistence()) {
+		if (file.exists() &&
+			(entity.isDeprecated() || !entity.hasPersistence())) {
+
 			System.out.println("Removing " + file);
 
 			file.delete();
